+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.1.0
+
+- installCertificateAuthority properly retrust root cert on macOS
+
 ## 2.0.0
 
 - requestCertificateForLocalhost changes